Corrected bridge weight markings for historical accuracy
Originally, I had put the "55" marking from the FV201 A45 on the front
and rear of the skin, not realizing that this was actually meant to
represent the vehicle's weight class historically - and the FV304 is a
20-ton vehicle, not the 55-ton sorry-excuse-for-a-heavy that is the A45
we currently have in-game. I've updated these markings accordingly to be
more historically-accurate.